refactor: Reorganize proto files to per-service structure
- Move common proto to pkg/proto/common/
- Move service protos to services/{service}/proto/ directories
- Update all import paths in Go files (24 files)
- Update proto generation script to use new paths
- Update go.mod files for all services
- Update BUILD.bazel files for new proto locations

This change improves operational manageability by organizing proto
files alongside their respective services, making it easier to
maintain and evolve each service independently.
