+++ pkg/proto/common/BUILD.bazel
@@ -9,7 +9,7 @@ proto_library(
 
 go_proto_library(
     name = "common_go_proto",
-    importpath = "github.com/Riku-KANO/kube-ec/proto/common",
+    importpath = "github.com/Riku-KANO/kube-ec/pkg/proto/common",
     proto = ":common_proto",
     visibility = ["//visibility:public"],
 )
