--- proto/go.mod
@@ -1,15 +1,0 @@
-module github.com/Riku-KANO/kube-ec/proto
-
-go 1.24.0
-
-require (
-	google.golang.org/grpc v1.76.0
-	google.golang.org/protobuf v1.36.6
-)
-
-require (
-	golang.org/x/net v0.42.0 // indirect
-	golang.org/x/sys v0.34.0 // indirect
-	golang.org/x/text v0.27.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20250804133106-a7a43d27e69b // indirect
-)
