+++ services/auth/internal/presentation/grpc/handler.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	pkgerrors "github.com/Riku-KANO/kube-ec/pkg/errors"
-	authpb "github.com/Riku-KANO/kube-ec/proto/auth"
-	commonpb "github.com/Riku-KANO/kube-ec/proto/common"
+	authpb "github.com/Riku-KANO/kube-ec/services/auth/proto"
+	commonpb "github.com/Riku-KANO/kube-ec/pkg/proto/common"
 	appauth "github.com/Riku-KANO/kube-ec/services/auth/internal/application/auth"
 )
