+++ services/auth/proto/BUILD.bazel
@@ -2,17 +2,17 @@ load("@rules_go//proto:def.bzl", "go_proto_library")
 load("@rules_proto//proto:defs.bzl", "proto_library")
 
 proto_library(
-    name = "user_proto",
-    srcs = ["user.proto"],
-    deps = ["//proto/common:common_proto"],
+    name = "auth_proto",
+    srcs = ["auth.proto"],
+    deps = ["//pkg/proto/common:common_proto"],
     visibility = ["//visibility:public"],
 )
 
 go_proto_library(
-    name = "user_go_proto",
+    name = "auth_go_proto",
     compilers = ["@rules_go//proto:go_grpc"],
-    importpath = "github.com/Riku-KANO/kube-ec/proto/user",
-    proto = ":user_proto",
+    importpath = "github.com/Riku-KANO/kube-ec/services/auth/proto",
+    proto = ":auth_proto",
     visibility = ["//visibility:public"],
-    deps = ["//proto/common:common_go_proto"],
+    deps = ["//pkg/proto/common:common_go_proto"],
 )
