+++ services/gateway/internal/infrastructure/grpc/helpers.go
@@ -7,7 +7,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/Riku-KANO/kube-ec/proto/common"
+	"github.com/Riku-KANO/kube-ec/pkg/proto/common"
 	"github.com/Riku-KANO/kube-ec/services/gateway/internal/domain/errors"
 )
