+++ services/gateway/internal/infrastructure/grpc/user_repository.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	userpb "github.com/Riku-KANO/kube-ec/proto/user"
+	userpb "github.com/Riku-KANO/kube-ec/services/user/proto"
 	"github.com/Riku-KANO/kube-ec/services/gateway/internal/domain/user"
 )
