+++ services/order/proto/BUILD.bazel
@@ -4,15 +4,15 @@ load("@rules_proto//proto:defs.bzl", "proto_library")
 proto_library(
     name = "order_proto",
     srcs = ["order.proto"],
-    deps = ["//proto/common:common_proto"],
+    deps = ["//pkg/proto/common:common_proto"],
     visibility = ["//visibility:public"],
 )
 
 go_proto_library(
     name = "order_go_proto",
     compilers = ["@rules_go//proto:go_grpc"],
-    importpath = "github.com/Riku-KANO/kube-ec/proto/order",
+    importpath = "github.com/Riku-KANO/kube-ec/services/order/proto",
     proto = ":order_proto",
     visibility = ["//visibility:public"],
-    deps = ["//proto/common:common_go_proto"],
+    deps = ["//pkg/proto/common:common_go_proto"],
 )
