+++ services/payment/proto/BUILD.bazel
@@ -4,15 +4,15 @@ load("@rules_proto//proto:defs.bzl", "proto_library")
 proto_library(
     name = "payment_proto",
     srcs = ["payment.proto"],
-    deps = ["//proto/common:common_proto"],
+    deps = ["//pkg/proto/common:common_proto"],
     visibility = ["//visibility:public"],
 )
 
 go_proto_library(
     name = "payment_go_proto",
     compilers = ["@rules_go//proto:go_grpc"],
-    importpath = "github.com/Riku-KANO/kube-ec/proto/payment",
+    importpath = "github.com/Riku-KANO/kube-ec/services/payment/proto",
     proto = ":payment_proto",
     visibility = ["//visibility:public"],
-    deps = ["//proto/common:common_go_proto"],
+    deps = ["//pkg/proto/common:common_go_proto"],
 )
