+++ services/product/go.mod
@@ -3,10 +3,10 @@ module github.com/Riku-KANO/kube-ec/services/product
 go 1.25
 
 require (
-	github.com/Riku-KANO/kube-ec/proto v0.0.0
 	github.com/google/uuid v1.6.0
 	github.com/lib/pq v1.10.9
 	google.golang.org/grpc v1.76.0
+	google.golang.org/protobuf v1.36.6
 )
 
 require (
@@ -16,5 +16,3 @@ require (
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20250804133106-a7a43d27e69b // indirect
 	google.golang.org/protobuf v1.36.6 // indirect
 )
-
-replace github.com/Riku-KANO/kube-ec/proto => ../../proto
