+++ services/product/main.go
@@ -7,7 +7,7 @@ import (
 	"net"
 	"os"
 
-	pb "github.com/Riku-KANO/kube-ec/proto/product"
+	pb "github.com/Riku-KANO/kube-ec/services/product/proto"
 	_ "github.com/lib/pq"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/reflection"
