+++ services/product/proto/BUILD.bazel
@@ -1,19 +1,18 @@
 load("@rules_go//proto:def.bzl", "go_proto_library")
 load("@rules_proto//proto:defs.bzl", "proto_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 proto_library(
     name = "product_proto",
     srcs = ["product.proto"],
-    deps = ["//proto/common:common_proto"],
+    deps = ["//pkg/proto/common:common_proto"],
     visibility = ["//visibility:public"],
 )
 
 go_proto_library(
     name = "product_go_proto",
     compilers = ["@rules_go//proto:go_grpc"],
-    importpath = "github.com/Riku-KANO/kube-ec/proto/product",
+    importpath = "github.com/Riku-KANO/kube-ec/services/product/proto",
     proto = ":product_proto",
     visibility = ["//visibility:public"],
-    deps = ["//proto/common:common_go_proto"],
+    deps = ["//pkg/proto/common:common_go_proto"],
 )
