+++ services/user/go.mod
@@ -3,9 +3,9 @@ module github.com/Riku-KANO/kube-ec/services/user
 go 1.25
 
 require (
-	github.com/Riku-KANO/kube-ec/proto v0.0.0
 	github.com/lib/pq v1.10.9
 	google.golang.org/grpc v1.76.0
+	google.golang.org/protobuf v1.36.6
 )
 
 require (
@@ -15,5 +15,3 @@ require (
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20250804133106-a7a43d27e69b // indirect
 	google.golang.org/protobuf v1.36.6 // indirect
 )
-
-replace github.com/Riku-KANO/kube-ec/proto => ../../proto
